+++ app/include/view/video_card.hpp
@@ -4,6 +4,8 @@
 #include <api/jellyfin/media.hpp>
 #include <utils/image.hpp>
 
+class SVGImage;
+
 class BaseVideoCard : public RecyclingGridItem {
 public:
     ~BaseVideoCard() { Image::cancel(this->picture); }
@@ -21,6 +23,7 @@ public:
 
     static VideoCardCell* create() { return new VideoCardCell(); }
 
+    BRLS_BIND(SVGImage, badgeTopRight, "video/card/badge/top");
     BRLS_BIND(brls::Label, labelTitle, "video/card/label/title");
     BRLS_BIND(brls::Label, labelExt, "video/card/label/ext");
     BRLS_BIND(brls::Label, labelRating, "video/card/label/rating");